refactor: T3 extract LSP param parsing and serialization to lsp_json
